Add board base module
